requirements.txt
pillow
screeninfo
mss
colour-science
numpy
pywizlight
pyyaml